Move justification to result package.
This package is a location of all the data that has been added from all
the data loaded in and should be separated from the regular data in lib/
